Route updated

+++ routes/web.php
@@ -52,13 +52,13 @@ Route::group(['middleware' => ['auth', 'checkUserType:admin']], function () {
     Route::put('/admin/patientlist/updated/{id}', [AdminPatientListController::class, 'updatedPatient'])->name('admin.updatedPatient');
     Route::delete('/admin/patientlist/delete/{id}', [AdminPatientListController::class, 'deletePatient'])->name('admin.deletePatient');
     // record
-    Route::get('/admin/patientlist/{patientlistId}/record', [AdminRecordController::class, 'showRecord'])->name('admin.showRecord');
-    Route::get('/admin/record/add/{patientlistId}', [AdminRecordController::class, 'createRecord'])->name('admin.record.create');
-    Route::post('/admin/records/store', [AdminRecordController::class, 'storeRecord'])->name('admin.record.store');
-    Route::get('/records/{patientlistId}/{recordId}/edit', [AdminRecordController::class, 'updateRecord'])->name('updateRecord');
-    Route::put('/admin/record/update/{patientlistId}/{recordId}', [AdminRecordController::class, 'updatedRecord'])->name('admin.record.update');
-    Route::delete('/records/{patientlist}/{record}', [AdminRecordController::class, 'deleteRecord'])->name('deleteRecord');
-    Route::get('/records/download/{record}', [AdminRecordController::class, 'downloadRecord'])->name('downloadRecord');
+    Route::get('/admin/patientlist/{patientlistId}/records', [AdminRecordController::class, 'showRecord'])->name('admin.showRecord');
+    Route::get('/admin/patientlist/{patientlistId}/record/add', [AdminRecordController::class, 'createRecord'])->name('admin.record.create');
+    Route::post('/admin/record/store', [AdminRecordController::class, 'storeRecord'])->name('admin.record.store');
+    Route::get('/admin/patientlist/{patientlistId}/record/update/{recordId}', [AdminRecordController::class, 'updateRecord'])->name('admin.updateRecord');
+    Route::put('/admin/patientlist/{patientlistId}/record//updated/{recordId}', [AdminRecordController::class, 'updatedRecord'])->name('admin.record.update');
+    Route::delete('/admin/records/{patientlist}/{record}', [AdminRecordController::class, 'deleteRecord'])->name('admin.deleteRecord');
+    Route::get('/admin/records/download/{record}', [AdminRecordController::class, 'downloadRecord'])->name('admin.downloadRecord');
 
 
     Route::get('/admin/search', [AdminPatientlistController::class, 'search'])->name('admin.search');
